# docs/swagger.yaml
basePath: /
definitions:
...
  dto.ChangePasswordDTO:
    properties:
      current_password:
        type: string
      new_password:
        type: string
    required:
    - current_password
    - new_password
    type: object
...
  dto.CreateResetTokenDTO:
    properties:
      email:
        type: string
    required:
    - email
    type: object
...
  dto.Login:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  dto.LoginResponse:
    properties:
      email:
        type: string
      message:
        type: string
    type: object
  dto.OTP:
    properties:
      email:
        type: string
      otp:
        type: string
    required:
    - email
    - otp
    type: object
  dto.OTPResponse:
    properties:
      email:
        type: string
      message:
        type: string
      role:
        type: string
      user_id:
        type: string
    type: object
  dto.ResetPasswordDTO:
    properties:
      new_password:
        type: string
      token:
        type: string
    required:
    - new_password
    - token
    type: object
  dto.Response:
    properties:
      message:
        type: string
    type: object
  dto.Verify:
    properties:
      password:
        type: string
      token:
        type: string
    required:
    - password
    - token
    type: object
host: e1ect.herokuapp.com
info:
  contact:
    email: surajnm15@gmail.com
    name: ELECT API Support
  description: This is the backend server of ELECT web application.
  title: ELECT REST API
  version: "1.0"
paths:
...
  /changepassword:
    post:
      operationId: changePassword
      parameters:
      - description: Change Password
        in: body
        name: changePassword
        required: true
        schema:
          $ref: '#/definitions/dto.ChangePasswordDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Change your password
      tags:
      - auth
  /createresettoken:
    post:
      operationId: createResetToken
      parameters:
      - description: Email
        in: body
        name: createResetToken
        required: true
        schema:
          $ref: '#/definitions/dto.CreateResetTokenDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Create a reset token and send email to reset password
      tags:
      - auth
  /login:
    post:
      operationId: login
      parameters:
      - description: User Login
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/dto.Login'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
      summary: User Login
      tags:
      - auth
  /otp:
    post:
      operationId: submitOTP
      parameters:
      - description: Verify OTP
        in: body
        name: otp
        required: true
        schema:
          $ref: '#/definitions/dto.OTP'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.OTPResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Submit OTP
      tags:
      - auth
  /refresh:
    post:
      description: A user needs a valid refresh token to access this endpoint.
      operationId: refresh
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Refresh Token
      tags:
      - auth
  /resetpassword:
    post:
      operationId: resetPassword
      parameters:
      - description: Reset Password
        in: body
        name: createResetToken
        required: true
        schema:
          $ref: '#/definitions/dto.ResetPasswordDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Reset password if you have a valid token
      tags:
      - auth
  /resettoken/{token}:
    post:
      operationId: checkResetTokenValidity
      parameters:
      - description: Reset Token
        in: path
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Check if reset token is valid or not
      tags:
      - auth
  /setpassword:
    post:
      operationId: verify
      parameters:
      - description: Verify
        in: body
        name: password
        required: true
        schema:
          $ref: '#/definitions/dto.Verify'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Verify Email and Set Password
      tags:
      - auth
  /ulogout:
    post:
      description: A user has to be logged in currently to access this endpoint.
      operationId: logout
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
      summary: User Logout
      tags:
      - auth
  /verifytoken/{token}:
    post:
      operationId: checkVerifyTokenValidity
      parameters:
      - description: Verify Token
        in: path
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Check if verify token is valid or not
      tags:
      - auth
swagger: "2.0"